// app/Utils/interface.ts
export interface footabll_stat {
    image: string;
    matches: number;
    wins: number;
    draws: number;
    losses: number;
    points: number;
    goals: number;
}


export interface cricket_stat {
    image: string;
    matches: number;
    wins: number;
    losses: number;
    points: number;
    runs: number;
    wickets: number;
}

export interface basketball_stat {
    image: string;
    matches: number;
    wins: number;
    losses: number;
    points: number;
    margin: number;
}

export interface volleyball_stat { 
    image: string;
    matches: number;
    wins: number;
    losses: number;
    points: number;
    margin: number;
}